final and constant.
